Add bilby pipe pe workflow

+++ gwcelery/tasks/inference.py
@@ -1,4 +1,4 @@
-"""Source Parameter Estimation with LALInference."""
+"""Source Parameter Estimation with LALInference and Bilby."""
 from distutils.spawn import find_executable
 from distutils.dir_util import mkpath
 import glob
@@ -22,7 +22,7 @@ from . import condor
 from . import gracedb
 
 
-ini_name = 'online_pe.ini'
+ini_name = 'online_lalinference_pe.ini'
 
 executables = {'datafind': 'gw_data_find',
                'mergeNSscript': 'lalinference_nest2pos',
@@ -142,7 +142,7 @@ def _find_appropriate_cal_env(trigtime, dir_name):
 @app.task(shared=False)
 def prepare_ini(frametype_dict, event, superevent_id=None):
     """Determine an appropriate PE settings for the target event and return ini
-    file content
+    file content for LALInference pipeline
     """
     # Get template of .ini file
     ini_template = env.get_template('online_pe.jinja2')
@@ -155,7 +155,9 @@ def prepare_ini(frametype_dict, event, superevent_id=None):
         'types': frametype_dict,
         'channels': app.conf['strain_channel_names'],
         'state_vector_channels': app.conf['state_vector_channel_names'],
-        'webdir': os.path.join(app.conf['pe_results_path'], event['graceid']),
+        'webdir': os.path.join(
+            app.conf['pe_results_path'], event['graceid'], 'lalinference'
+        ),
         'paths': [{'name': name, 'path': find_executable(executable)}
                   for name, executable in executables.items()],
         'h1_calibration': _find_appropriate_cal_env(
@@ -176,12 +178,12 @@ def prepare_ini(frametype_dict, event, superevent_id=None):
     ini_rota = ini_template.render(ini_settings)
     ini_settings.update({'use_of_ini': 'online'})
     ini_online = ini_template.render(ini_settings)
-    # upload ini file to GraceDB
+    # upload LALInference ini file to GraceDB
     if superevent_id is not None:
         gracedb.upload.delay(
             ini_rota, filename=ini_name, graceid=superevent_id,
-            message='Automatically generated LALInference configuration file'
-                    ' for this event.',
+            message=('Automatically generated LALInference configuration file'
+                     ' for this event.'),
             tags='pe')
 
     return ini_online
@@ -195,58 +197,108 @@ def pre_pe_tasks(event, superevent_id):
 
 
 @app.task(shared=False)
-def dag_prepare(
-    coinc_psd, ini_contents, rundir, superevent_id
-):
-    """Create a Condor DAG to run LALInference on a given event.
+def dag_prepare(coinc_psd, rundir, superevent_id, preferred_event_id,
+                pe_pipeline, ini_contents=None):
+    """Create a Condor DAG to run Parameter Estimation on a given event.
 
     Parameters
     ----------
-    coinc_psd : tuple
-        The tuple of the byte contents of ``coinc.xml`` and ``psd.xml.gz``
-    ini_contents : str
-        The content of online_pe.ini
+    coinc_psd : tuple or json contents
+        For lalinference, tuple of the byte contents of
+        ``coinc.xml`` and ``psd.xml.gz``
+        For bilby, json contents retrieved from gracedb.get_event()
     rundir : str
         The path to a run directory where the DAG file exits
     superevent_id : str
         The GraceDB ID of a target superevent
+    preferred_event_id : str
+        The GraceDB ID of a target preferred event
+    pe_pipeline : str
+        The parameter estimation pipeline used
+        Either lalinference OR bilby
+    ini_contents : str
+        The content of online_lalinference_pe.ini
 
     Returns
     -------
     submit_file : str
         The path to the .sub file
     """
-    coinc_contents, psd_contents = coinc_psd
 
-    # write down coicn.xml in the run directory
-    path_to_coinc = os.path.join(rundir, 'coinc.xml')
-    with open(path_to_coinc, 'wb') as f:
-        f.write(coinc_contents)
+    if pe_pipeline == 'lalinference':
 
-    # write down psd.xml.gz
-    if psd_contents is not None:
-        path_to_psd = os.path.join(rundir, 'psd.xml.gz')
-        with open(path_to_psd, 'wb') as f:
-            f.write(psd_contents)
-        psd_arg = ['--psd', path_to_psd]
-    else:
-        psd_arg = []
+        coinc_contents, psd_contents = coinc_psd
 
-    # write down .ini file in the run directory.
-    path_to_ini = rundir + '/' + ini_name
-    with open(path_to_ini, 'w') as f:
-        f.write(ini_contents)
+        # write down coinc.xml in the run directory
+        path_to_coinc = os.path.join(rundir, 'coinc.xml')
+        with open(path_to_coinc, 'wb') as f:
+            f.write(coinc_contents)
 
-    # run lalinference_pipe
+        # write down psd.xml.gz
+        if psd_contents is not None:
+            path_to_psd = os.path.join(rundir, 'psd.xml.gz')
+            with open(path_to_psd, 'wb') as f:
+                f.write(psd_contents)
+            psd_arg = '--psd {}'.format(path_to_psd)
+        else:
+            psd_arg = ''
+
+        # write down .ini file in the run directory.
+        path_to_ini = rundir + '/' + ini_name
+        with open(path_to_ini, 'w') as f:
+            f.write(ini_contents)
+
+        setup_arg = 'lalinference_pipe --run-path {} '.format(rundir) +\
+                    '--coinc {} '.format(path_to_coinc) +\
+                    '{} {}'.format(path_to_ini, psd_arg)
+
+        dag_arg = 'condor_submit_dag -no_submit {}/multidag.dag'.format(rundir)
+
+    if pe_pipeline == 'bilby':
+
+        path_to_json = os.path.join(rundir, 'event.json')
+        with open(path_to_json, 'w') as f:
+            json.dump(coinc_psd, f, indent=2)
+
+        path_to_webdir = os.path.join(
+            app.conf['pe_results_path'], preferred_event_id, 'bilby'
+        )
+
+        # For bilby_pipe PE ligo-py36 conda env needs to be sourced.
+        if app.conf['sentry_environment'] == 'production':
+            setup_arg = 'source /cvmfs/ligo-containers.opensciencegrid.org' +\
+                        '/lscsoft/conda/latest/etc/profile.d/conda.sh ' +\
+                        '&& conda activate ligo-py36 && bilby_pipe_gracedb ' +\
+                        '--webdir {} '.format(path_to_webdir) +\
+                        '--outdir {} --json {}'.format(rundir, path_to_json)
+
+        else:
+            setup_arg = 'source /cvmfs/ligo-containers.opensciencegrid.org' +\
+                        '/lscsoft/conda/latest/etc/profile.d/conda.sh ' +\
+                        '&& conda activate ligo-py36 && bilby_pipe_gracedb ' +\
+                        '--channel-dict o2replay ' +\
+                        '--webdir {} '.format(path_to_webdir) +\
+                        '--outdir {} --json {}'.format(rundir, path_to_json)
     try:
-        lalinference_arg = ['lalinference_pipe', '--run-path', rundir,
-                            '--coinc', path_to_coinc, path_to_ini] + psd_arg
-        subprocess.run(lalinference_arg, stdout=subprocess.PIPE,
-                       stderr=subprocess.PIPE, check=True)
-        subprocess.run(['condor_submit_dag', '-no_submit',
-                        rundir + '/multidag.dag'],
-                       stdout=subprocess.PIPE, stderr=subprocess.PIPE,
-                       check=True)
+        subprocess.run(
+            setup_arg, stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE, check=True, shell=True
+        )
+        if pe_pipeline == 'bilby':
+
+            # Uploads bilby ini file to GraceDB
+            _upload_result(rundir, 'bilby_config.ini', superevent_id,
+                           'Automatically generated Bilby configuration file',
+                           'pe', 'online_bilby_pe.ini').delay()
+
+            path_to_dag = glob.glob(rundir + '/submit/dag*.submit')
+            dag_arg = 'condor_submit_dag -no_submit {}'.format(path_to_dag[0])
+
+        subprocess.run(
+            dag_arg, stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE, check=True, shell=True
+        )
+
     except subprocess.CalledProcessError as e:
         contents = b'args:\n' + json.dumps(e.args[1]).encode('utf-8') + \
                    b'\n\nstdout:\n' + e.stdout + b'\n\nstderr:\n' + e.stderr
@@ -260,9 +312,15 @@ def dag_prepare(
     finally:
         # Remove the ini file so that people do not accidentally use this ini
         # file and hence online-PE-only nodes.
-        os.remove(path_to_ini)
+        if pe_pipeline == 'lalinference':
+            os.remove(path_to_ini)
 
-    return rundir + '/multidag.dag.condor.sub'
+    if pe_pipeline == 'lalinference':
+        path_to_sub = rundir + '/multidag.dag.condor.sub'
+    if pe_pipeline == 'bilby':
+        sub = glob.glob(rundir + '/submit/dag*.submit.condor.sub')
+        path_to_sub = sub[0]
+    return path_to_sub
 
 
 def _find_paths_from_name(directory, name):
@@ -334,9 +392,17 @@ def job_error_notification(request, exc, traceback, superevent_id, rundir):
 
 
 @app.task(ignore_result=True, shared=False)
-def _upload_url(pe_results_path, graceid):
+def _upload_url(pe_results_path, graceid, pe_pipeline):
     """Upload url of a page containing all of the plots."""
-    path_to_posplots, = _find_paths_from_name(pe_results_path, 'posplots.html')
+    if pe_pipeline == 'lalinference':
+        path_to_posplots, = _find_paths_from_name(
+            pe_results_path, 'posplots.html'
+        )
+    elif pe_pipeline == 'bilby':
+        path_to_posplots, = _find_paths_from_name(
+            pe_results_path, 'home.html'
+        )
+
     baseurl = urllib.parse.urljoin(
                   app.conf['pe_results_url'],
                   os.path.relpath(
@@ -346,8 +412,8 @@ def _upload_url(pe_results_path, graceid):
               )
     gracedb.upload.delay(
         filecontents=None, filename=None, graceid=graceid,
-        message=('LALInference online parameter estimation finished.'
-                 '<a href={}>results</a>').format(baseurl),
+        message=('Online {} parameter estimation finished.'
+                 '<a href={}>results</a>').format(pe_pipeline, baseurl),
         tags='pe'
     )
 
@@ -387,7 +453,7 @@ def clean_up(rundir):
 
 
 @app.task(ignore_result=True, shared=False)
-def dag_finished(rundir, preferred_event_id, superevent_id):
+def dag_finished(rundir, preferred_event_id, superevent_id, pe_pipeline):
     """Upload PE results and clean up run directory
 
     Parameters
@@ -398,39 +464,76 @@ def dag_finished(rundir, preferred_event_id, superevent_id):
         The GraceDB ID of a target preferred event
     superevent_id : str
         The GraceDB ID of a target superevent
+    pe_pipeline : str
+        The parameter estimation pipeline used
+        Either lalinference OR bilby
 
     Returns
     -------
     tasks : canvas
         The work-flow for uploading PE results
     """
-    # get path to pe results
-    pe_results_path = \
-        os.path.join(app.conf['pe_results_path'], preferred_event_id)
 
-    # FIXME: _upload_url.si has to be out of group for gracedb.create_label.si
-    # to run
+    if pe_pipeline == 'lalinference':
+        # path to lalinference pe results
+        pe_results_path = os.path.join(
+            app.conf['pe_results_path'], preferred_event_id, 'lalinference'
+        )
+
+        uploads = [
+            (rundir, 'posterior*.hdf5',
+             'LALInference posterior samples',
+             'LALInference.posterior_samples.hdf5'),
+            (pe_results_path, 'extrinsic.png',
+             'LALInference corner plot for extrinsic parameters',
+             'LALInference.extrinsic.png'),
+            (pe_results_path, 'sourceFrame.png',
+             'LALInference corner plot for source frame parameters',
+             'LALInference.intrinsic.png'),
+        ]
+
+    elif pe_pipeline == 'bilby':
+        # path to bilby pe results
+        pe_results_path = os.path.join(
+            app.conf['pe_results_path'], preferred_event_id, 'bilby'
+        )
+
+        resultdir = rundir + '/result'
+
+        resultfile = preferred_event_id + '_combined_result.json'
+        intrinsic = preferred_event_id + '_combined_intrinsic_corner.png'
+        extrinsic = preferred_event_id + '_combined_extrinsic_corner.png'
+
+        uploads = [
+            (resultdir, resultfile,
+             'Bilby posterior samples',
+             'Bilby.posterior_samples.json'),
+            (resultdir, extrinsic,
+             'Bilby corner plot for extrinsic parameters',
+             'Bilby.extrinsic.png'),
+            (resultdir, intrinsic,
+             'Bilby corner plot for intrinsic parameters',
+             'Bilby.intrinsic.png'),
+        ]
+
+    upload_tasks = [
+        _upload_result(
+            dir, name1, superevent_id, comment, 'pe', name2
+        ) for dir, name1, comment, name2 in uploads
+    ]
+
+    # FIXME: _upload_url.si has to be out of group for
+    # gracedb.create_label.si to run
     (
-        _upload_url.si(pe_results_path, superevent_id) |
-        group(
-            _upload_result(
-                rundir, 'posterior*.hdf5', superevent_id,
-                'LALInference posterior samples', 'pe',
-                'LALInference.posterior_samples.hdf5'
-            ),
-            _upload_result(
-                pe_results_path, 'extrinsic.png', superevent_id,
-                'Corner plot for extrinsic parameters', 'pe',
-                'LALInference.extrinsic.png'
-            ),
-            _upload_result(
-                pe_results_path, 'sourceFrame.png', superevent_id,
-                'Corner plot for source frame parameters', 'pe',
-                'LALInference.intrinsic.png'
-            )
-        ) | gracedb.create_label.si('PE_READY', superevent_id) |
+        _upload_url.si(pe_results_path, superevent_id, pe_pipeline)
+        |
+        group(upload_tasks)
+        |
         clean_up.si(rundir)
     ).delay()
+
+    if pe_pipeline == 'lalinference':
+        gracedb.create_label.delay('PE_READY', superevent_id)
 
 
 @gracedb.task(shared=False)
@@ -445,45 +548,64 @@ def _download_psd(gid):
 
 
 @app.task(ignore_result=True, shared=False)
-def start_pe(ini_contents, preferred_event_id, superevent_id):
-    """Run LALInference on a given event.
+def start_pe(ini_contents, preferred_event_id, superevent_id, pe_pipeline):
+    """Run Parameter Estimation on a given event.
 
     Parameters
     ----------
     ini_contents : str
-        The content of online_pe.ini
+        The content of online_lalinference_pe.ini
     preferred_event_id : str
         The GraceDB ID of a target preferred event
     superevent_id : str
         The GraceDB ID of a target superevent
+    pe_pipeline : str
+        The parameter estimation pipeline used
+        lalinference OR bilby
+
     """
+
     gracedb.upload.delay(
         filecontents=None, filename=None, graceid=superevent_id,
-        message=('starting LALInference online parameter estimation '
-                 'for {}').format(preferred_event_id),
+        message=('Starting {} online parameter estimation '
+                 'for {}').format(pe_pipeline, preferred_event_id),
         tags='pe'
     )
 
     # make a run directory
-    lalinference_dir = os.path.expanduser('~/.cache/lalinference')
-    mkpath(lalinference_dir)
-    rundir = tempfile.mkdtemp(dir=lalinference_dir,
-                              prefix='{}_'.format(superevent_id))
+    pipeline_dir = os.path.expanduser('~/.cache/{}'.format(pe_pipeline))
+    mkpath(pipeline_dir)
+    rundir = tempfile.mkdtemp(
+        dir=pipeline_dir, prefix='{}_'.format(superevent_id)
+    )
+
     # give permissions to read the files under the run directory so that PE
     # ROTA people can check the status of parameter estimation.
     os.chmod(rundir, 0o755)
 
-    (
-        ordered_group(
+    if pe_pipeline == 'lalinference':
+        canvas = ordered_group(
             gracedb.download.s('coinc.xml', preferred_event_id),
             _download_psd.s(preferred_event_id)
         )
+
+    elif pe_pipeline == 'bilby':
+        canvas = gracedb.get_event.si(preferred_event_id)
+        # ini contents not passed to bilby
+        ini_contents = None
+
+    canvas |= (
+        dag_prepare.s(
+            rundir, superevent_id, preferred_event_id,
+            pe_pipeline, ini_contents
+        )
         |
-        dag_prepare.s(ini_contents, rundir, superevent_id)
-        |
-        condor.submit.s().on_error(
+        condor.submit.s(pe_pipeline).on_error(
             job_error_notification.s(superevent_id, rundir)
         )
         |
-        dag_finished.si(rundir, preferred_event_id, superevent_id)
-    ).delay()
+        dag_finished.si(
+            rundir, preferred_event_id, superevent_id, pe_pipeline
+        )
+    )
+    canvas.delay()
